Move (un)mapping rules into a separate module
